+++ NEWS.md
@@ -2,3 +2,6 @@
 
 * Added exported package data variable `signature`, `etiology`, `COSMIC_v3.2`,
 `COSMIC_v3.1` and `COSMIC_v3.0`
+
+* Added exported function `get_etiology`, `SBS96_ID_to_SBS192_ID` and automated
+tests